refactor(invocation): modularize invocation handling into manager, service, and tracker with improved caching, state tracking, and structured error handling
